fix(sdk): package net8 live contract surface
