+++ dotnet/live-contracts/Devolutions.MultiPwsh.LiveContracts.csproj
@@ -1,0 +1,9 @@
+<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+    <AssemblyName>Devolutions.MultiPwsh.LiveContracts</AssemblyName>
+    <RootNamespace>Devolutions.PowerShell.Ffi.LiveObjects</RootNamespace>
+    <LangVersion>latest</LangVersion>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+</Project>
