--- dotnet/sdk-ffi/buildTransitive/Devolutions.MultiPwsh.LiveContract.Generator.targets
@@ -1,6 +1,0 @@
-<Project>
-  <ItemGroup>
-    <Analyzer Include="$(MSBuildThisFileDirectory)..\analyzers\dotnet\cs\Devolutions.MultiPwsh.LiveContract.Generator.dll" />
-    <CompilerVisibleProperty Include="LiveContractMode" />
-  </ItemGroup>
-</Project>
